Refactor TenantSeedCommand to use correct DatabaseSeeder reference; add seeders for Configuration, CanchasYHorarios, and EquiposYJugadores with initial data population.
