+++ database/seeders/CanchasYHorariosSeeder.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Database\Seeders\Complejos;
+namespace Database\Seeders;
 
 use Illuminate\Database\Seeder;
 use App\Models\Cancha;
