+++ database/seeders/ConfigurationSeeder.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Database\Seeders\Complejos;
+namespace Database\Seeders;
 
 use Illuminate\Database\Console\Seeds\WithoutModelEvents;
 use Illuminate\Database\Seeder;
