+++ database/seeders/DatabaseSeeder.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Database\Seeders\Complejos;
+namespace Database\Seeders;
 
 use App\Models\User;
 // use Illuminate\Database\Console\Seeds\WithoutModelEvents;
